backend/pyproject.toml
[project]
name = "qualify"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncssh>=2.14.0",
    "pydantic>=2.7.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "pyyaml>=6.0.0",
    "keyring>=25.0.0",
    "uuid-utils>=0.9.0",
]

[project.optional-dependencies]
dev = ["pyinstaller>=6.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/qualify"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
